backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
PyMuPDF
pillow
pydantic
python-dotenv
numpy